Agreements/requirements.txt
python-docx>=1.2.0

